Changes to files_io
Some minor changes to progress bar.

I had to move the create_inscopix_projects function to filehandler.py because there is circular import loop between filehandler and files_io. Filehandler imports files_io while files_io imports filehandler
